% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_list}
\alias{check_list}
\title{Check if all required variables are a list}
\usage{
check_list(list_to_check, required_variables)
}
\arguments{
\item{list_to_check}{A list.}

\item{required_variables}{A vector containing the names of the variables which must be in \code{list_to_check}.}
}
\value{
It returns \code{TRUE} if \code{list_to_check} contains all the required variables specified in \code{required_variables}.
}
\description{
Check if all required variables are a list
}
