% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributions.R
\name{contributions}
\alias{contributions}
\title{Contributions}
\usage{
contributions(out_MCMC, reference = NULL, chains = FALSE, verbose = TRUE)
}
\arguments{
\item{out_MCMC}{A list containing the results of an Adaptive Gibbs Sampler obtained using the function \code{\link{cosin}}.}

\item{reference}{An integer specifying the reference iteration used to sort the contributions of each iteration; if \code{reference=NULL}, the contributions are sorted using the Frobenius norm. Default is \code{NULL}.}

\item{chains}{Logical: if \code{TRUE}, the function returns also the draws from the posterior distribution of \eqn{C^*_1,\ldots,C^*_{k_{max}}}.}

\item{verbose}{Logical: if \code{TRUE}, print the number of active factors every 50 iterations. Default is \code{TRUE}.}
}
\value{
A list with the following elements:
\itemize{
\item \code{C1}: a list of matrices containing the posterior means of \eqn{C^*_1,\ldots,C^*_{k_{max}}}.
\item \code{C2}: a list of matrices containing the posterior means of the square of the elements of \eqn{C^*_1,\ldots,C^*_{k_{max}}}.
\item \code{reference}: index of the iteration used as reference.
\item \code{chains}: a list containing draws from the posterior distribution of \eqn{C^*_1,\ldots,C^*_{k_{max}}}.
}
}
\description{
Compute the sorted contributions \eqn{C^*_1,\ldots,C^*_{k_{max}}} for each iteration.
}
\seealso{
This function is applied to an output of \code{\link{cosin}}.
}
