% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lposterior.R
\name{det_log_omega}
\alias{det_log_omega}
\title{Logarithm of the determinant of a matrix}
\usage{
det_log_omega(lambda, sigmai)
}
\arguments{
\item{lambda}{A pxk matrix of factorial weights, \eqn{\Lambda}.}

\item{sigmai}{A p-dimensional vector containing \eqn{(\sigma^2_1,\ldots,\sigma^2_p)}.}
}
\value{
A number.
}
\description{
Compute the logarithm of the determinant of \eqn{\Omega} using the following lemma
\deqn{\det(A+UV^\top) = \det(I_m+V^\top A^{-1}U)\det(A)}
where \eqn{A} is a \eqn{n\times n} matrix and \eqn{U}, \eqn{V} are \eqn{n\times m} matrices.
Since \eqn{\Omega=\Lambda\Lambda^\top+\Sigma}, \eqn{\det(\Omega)} can be written as follows
\deqn{\det({\Omega}) = \det(\Lambda\Lambda^\top+\Sigma) = \det(I_k+\Lambda^\top\Sigma^{-1}\Lambda)\det(\Sigma)}
where \eqn{\Lambda} is a \eqn{p\times k} matrix and \eqn{\Sigma=\text{diag}(\sigma^2_1,\ldots,\sigma^2_p)} is a \eqn{p\times p} diagonal matrix.
Then, \eqn{\log(\det(\Omega))} can be written as follows
\deqn{\log(\det(\Omega)) = \log(\det(I_k+\Lambda^\top\Sigma^{-1}\Lambda))+\log(\sigma^2_1)+\ldots+\log(\sigma^2_p)}
}
\seealso{
This function is used in \code{\link{lposterior_function}}.
}
