% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lposterior.R
\name{gen_z}
\alias{gen_z}
\title{Generating values from a truncated normal distribution}
\usage{
gen_z(Y, mean, ps, y_max)
}
\arguments{
\item{Y}{A nxp matrix.}

\item{mean}{A nxp matrix containing the means of the truncated normal distribution.}

\item{ps}{DA SCRIVERE}

\item{y_max}{A number specifying the maximum value an element of \code{Y} can take. \code{Inf} is allowed.}
}
\value{
A nxp matrix.
}
\description{
Generate values from a truncated normal distribution.
}
\seealso{
This function is used in \code{\link{lposterior_function}}.
}
