% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lposterior.R
\name{log_multinorm}
\alias{log_multinorm}
\title{Logarithm of a multivariate normal density}
\usage{
log_multinorm(z, mu, omega_inv, omega_log_det)
}
\arguments{
\item{z}{A vector of values on which density is to be calculated.}

\item{mu}{A vector of the means of the multivariate normal distribution.}

\item{omega_inv}{Inverse of the variance matrix of the multivariate normal distribution.}

\item{omega_log_det}{Logarithm of the determinant of the variance matrix of the multivariate normal distribution.}
}
\value{
A vector with the same dimensions as \code{z}.
}
\description{
Compute the logarithm of a multivariate normal density.
}
\seealso{
This function is used in \code{\link{lposterior_function}}.
}
