% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lposterior.R
\name{lposterior_function}
\alias{lposterior_function}
\title{Log-posterior probability of a single MCMC iteration}
\usage{
lposterior_function(ind, out_MCMC, columns = "k")
}
\arguments{
\item{ind}{An integer specifying the index of the iteration of interest.}

\item{out_MCMC}{A list containing the results of an Adaptive Gibbs Sampler obtained using the function \code{\link{cosin}}.}

\item{columns}{a string specifying whether to consider only active factors (\code{kstar}) or both active and inactive factors (\code{k}). Default is \code{"k"}.}
}
\value{
A number which is the log-posterior probability of the \code{ind}th iteration of a MCMC chain.
}
\description{
Compute the log-posterior probability of a single MCMC iteration.
}
\seealso{
This function is applied to an output of \code{\link{cosin}}. This function is used in \code{\link{lposterior}}.
}
