% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{omega_inversion}
\alias{omega_inversion}
\title{Inversion of omega}
\usage{
omega_inversion(lambda, sigmai)
}
\arguments{
\item{lambda}{A pxk matrix of factorial weights.}

\item{sigmai}{A p-dimensional vector of \code{1/sigma^2}.}
}
\value{
A pxp matrix.
}
\description{
This function is called in 'posterior_mean' and 'lposterior_function'.
}
