% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_mean.R
\name{posterior_mean}
\alias{posterior_mean}
\title{Posterior mean}
\usage{
posterior_mean(out_MCMC, parameters = "all", columns = "k")
}
\arguments{
\item{out_MCMC}{A list containing the results of an Adaptive Gibbs Sampler obtained using the function \code{\link{cosin}}.}

\item{parameters}{A vector containing the names of the parameters for which you want to compute the posterior mean. The possible valid strings are \code{"beta"}, \code{"gammaT"}, \code{"sigmacol"}, \code{"omega"}, \code{"omega_inv"} and \code{"omega_pcorr"}. Default is \code{"all"}, which is equivalent to writing \code{c("beta", "gammaT", "sigmacol", "omega", "omega_inv", "omega_pcorr")}.}

\item{columns}{A string specifying whether to consider only active factors (\code{"kstar"}) or both active and inactive factors (\code{"k"}) for the computation of \code{omega}, \code{omega_inv} and \code{"omega_pcorr"}. Default is \code{"k"}.}
}
\value{
A list containing the posterior means of the parameters specified in \code{parameters}.
}
\description{
Compute the posterior mean of one or more parameters.
}
\seealso{
This function is applied to an output of \code{\link{cosin}}.
}
