% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{rescale_parameter}
\alias{rescale_parameter}
\title{Rescale one or more parameters}
\usage{
rescale_parameter(out_MCMC, parameters = "all", columns = "k", samples = NULL)
}
\arguments{
\item{out_MCMC}{A list containing the results of an Adaptive Gibbs Sampler obtained using the function \code{\link{cosin}}.}

\item{parameters}{A vector of strings specifying the parameters for which you want to compute the posterior mean. The possible valid strings are \code{gammaB}, \code{eta} and \code{lambda}. Default is \code{all}, which is equivalent to writing \code{c("gammaB", "eta", "lambda")}.}

\item{columns}{A string specifying whether to consider only active factors (\code{kstar}) or both active and inactive factors (\code{k}). Default is \code{k}.}

\item{samples}{A vector of integers specifying the iterations to consider. By default, all iterations are considered.}
}
\value{
A list containing the MCMC chains of the parameters specified in \code{parameters}.
}
\description{
Add columns to one or more parameters whose number of columns depends on the number of factors so that the parameters of all MCMC iterations have the same number of columns. This number can be equal to the maximum number of active factors or the maximum number of factors, active and inactive, observed during the Adaptive Gibbs Sampler.
}
\seealso{
This function is applied to an output of \code{\link{cosin}}.
}
