% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wB}
\alias{wB}
\title{Gene-specific biological meta-covariates}
\format{
A data frame with 100 rows and 3 columns:
\itemize{
  \item \code{Alzheimer_disease}
  \item \code{Parkinson_disease}
  \item \code{Huntington_disease}
}
}
\usage{
wB
}
\description{
A \eqn{p\times q_B} matrix \eqn{x} of gene-specific biological meta-covariates.
}
\keyword{datasets}
