% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wT}
\alias{wT}
\title{Gene-specific technical meta-covariates}
\format{
A data frame with 100 rows and 2 columns:
\itemize{
  \item \code{length}
  \item \code{gc_content}
}
}
\usage{
wT
}
\description{
A \eqn{p\times q_T} matrix of gene-specific technical meta-covariates.
}
\keyword{datasets}
