% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x}
\alias{x}
\title{Cell-specific covariates}
\format{
A data frame with 20 rows and 4 columns:
\itemize{
  \item \code{unaligned}
  \item \code{aligned_unmapped}
  \item \code{number_of_genes}
  \item \code{total_count_per_cell}
}
}
\usage{
x
}
\description{
A \eqn{n\times d} matrix of cell-specific covariates.
}
\keyword{datasets}
