% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{y}
\alias{y}
\title{Matrix of gene expression}
\format{
A matrix with 20 rows and 100 columns.
}
\usage{
y
}
\description{
A \eqn{n\times p} matrix \eqn{y} of gene expression containing the counts
for \eqn{p} genes measured in on \eqn{n} cells.
}
\keyword{datasets}
